requirements.txt
annotated-types==0.7.0
anyio==4.9.0
async-timeout==5.0.1
click==8.2.1
exceptiongroup==1.3.0
fastapi==0.115.12
h11==0.16.0
idna==3.10
joblib==1.5.1
jpype1==1.5.2
lxml==5.4.0
mysql-connector-python==9.3.0
numpy==2.2.6
packaging==25.0
pandas==2.2.3
psutil==7.0.0
pydantic==2.11.5
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
redis==6.2.0
scikit-learn==1.6.1
scipy==1.15.3
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
structlog==25.4.0
threadpoolctl==3.6.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
uvicorn==0.34.3
pydantic_settings==2.9.1
pydantic_extra_types==2.10.5
loguru==0.7.3
msgspec==0.19.0
asgiref==3.8.1
opensearch-dsl==2.1.0
opensearch-py==2.8.0
packaging==25.0
aiohttp==3.12.9
httpx==0.28.1
psutil==7.0.0
torch --index-url https://download.pytorch.org/whl/cpu
sentence-transformers
sqlalchemy